requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
fastapi==0.120.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
pillow==12.0.0
pydantic==2.12.3
pydantic-core==2.41.4
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
sqlalchemy==2.0.44
starlette==0.48.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.38.0
